Put spherical coordinate transforms in a single file.
Update definitions in the NEW file.
